+++ server/events/events.go
@@ -49,14 +49,14 @@ type TunnelRequested struct {
 type TunnelStarted struct {
 	Host           string   `json:"host"`
 	Protocol       Protocol `json:"protocol"`
-	PublicServer   int      `json:"public_server"`
-	PrivateServer  int      `json:"private_server"`
+	PublicServer   uint16   `json:"public_server"`
+	PrivateServer  uint16   `json:"private_server"`
 	UserMessage    string   `json:"user_message"`
-	MaxConnections int      `json:"max_connections"`
+	MaxConnections uint16   `json:"max_connections"`
 }
 
 type ConnectionReceived struct {
 	ClientIP    string `json:"client_ip"`
-	ClientPort  int    `json:"client_port"`
+	ClientPort  uint16 `json:"client_port"`
 	RateLimited bool   `json:"rate_limited"`
 }
